--- EnumAxis
@@ -1,7 +1,0 @@
-export { Axis };
-
-enum Axis {
-  X,
-  Y,
-  Z,
-}
